T19: rename and moved package version configuration. Deleted unnecessary projects...
