+++ NugetAnalyzer.Common/NugetAnalyzer.Common.csproj
@@ -4,5 +4,9 @@
     <TargetFramework>netcoreapp2.2</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\NugetAnalyzer.Dtos\NugetAnalyzer.Dtos.csproj" />
+  </ItemGroup>
+
 
 </Project>
