--- NugetAnalyzer.Web.Infrastructure/Configurations/GitHubEndPointsConfigurations.cs
@@ -1,11 +1,0 @@
-﻿namespace NugetAnalyzer.Web.Infrastructure.Configurations
-{
-    public class GitHubEndPointsConfigurations
-    {
-        public string AuthorizationEndpoint { get; set; }
-
-        public string TokenEndpoint { get; set; }
-
-        public string UserInformationEndpoint { get; set; }
-    }
-}
