--- NugetAnalyzer.Web.Infrastructure/Configurations/GitHubSecretsConfigurations.cs
@@ -1,11 +1,0 @@
-﻿namespace NugetAnalyzer.Web.Infrastructure.Configurations
-{
-	public class GitHubSecretsConfigurations
-	{
-		public string ClientId { get; set; }
-
-		public string ClientSecret { get; set; }
-
-		public string CallbackPath { get; set; }
-	}
-}
