--- NugetAnalyzer.Web.Infrastructure/Extensions/ApplicationBuilderExtensions.cs
@@ -1,25 +1,0 @@
-﻿using System.Globalization;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Localization;
-
-namespace NugetAnalyzer.Web.Infrastructure.Extensions
-{
-    public static class LocalizationExtensions
-    {
-        public static void UseConfiguredLocalization(this IApplicationBuilder app)
-        {
-            var supportedCultures = new[]
-            {
-                new CultureInfo("ru"),
-                new CultureInfo("en")
-            };
-
-            app.UseRequestLocalization(new RequestLocalizationOptions
-            {
-                DefaultRequestCulture = new RequestCulture("en"),
-                SupportedCultures = supportedCultures,
-                SupportedUICultures = supportedCultures
-            });
-        }
-    }
-}
