--- NugetAnalyzer.Web.Infrastructure2/NugetAnalyzer.Web.Infrastructure2.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Localization" Version="2.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\NugetAnalyzer.Dtos\NugetAnalyzer.Dtos.csproj" />
-  </ItemGroup>
-
-</Project>
