--- NugetAnalyzer.Web.Infrastructure2/Options/GitHubSecretsOptions.cs
@@ -1,11 +1,0 @@
-﻿namespace NugetAnalyzer.Web.Options
-{
-    public class GitHubSecretsOptions
-    {
-        public string ClientId { get; set; }
-
-        public string ClientSecret { get; set; }
-
-        public string CallbackPath { get; set; }
-    }
-}
