+++ NugetAnalyzer.Web/Controllers/HomeController.cs
@@ -1,19 +1,11 @@
-﻿using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using NugetAnalyzer.BLL.Interfaces;
+﻿using Microsoft.AspNetCore.Mvc;
 
 namespace NugetAnalyzer.Web.Controllers
 {
     public class HomeController : Controller
     {
-        private readonly IRepositoryService repositoryService;
-        public HomeController(IRepositoryService repositoryService)
+        public IActionResult Index()
         {
-            this.repositoryService = repositoryService;
-        }
-        public async Task<IActionResult> Index()
-        {
-            var result = await repositoryService.GetAnalyzedRepositoriesAsync(repository => repository.UserId == 1);
             return View();
         }
     }
