+++ NugetAnalyzer.sln
@@ -15,7 +15,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NugetAnalyzer.BLL.Test", "N
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NugetAnalyzer.Common", "NugetAnalyzer.Common\NugetAnalyzer.Common.csproj", "{A2CA0EA5-6FA8-4ACB-9FE4-3ECCA4B3B46E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NugetAnalyzer.Dtos", "NugetAnalyzer.Dtos\NugetAnalyzer.Dtos.csproj", "{22465B5D-EDB5-4196-8309-A404DB534914}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NugetAnalyzer.Dtos", "NugetAnalyzer.Dtos\NugetAnalyzer.Dtos.csproj", "{22465B5D-EDB5-4196-8309-A404DB534914}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
